add get one planet test

+++ app/models/planet.py
@@ -4,7 +4,7 @@ class Planet(db.Model):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     name: Mapped[str]
     description: Mapped[str] = mapped_column(nullable=True)
-    radius_in_mi: Mapped[int]
+    radius_in_mi: Mapped[float]
 
     def to_dict(self):
         return {
